// src/app/enums/EROUTES.ts
export enum EROUTES {
    // Defaults
    BASE = '',
    SEPARATOR = '/',

    // General
    ABOUT_PAGE = 'about-page',

    // Products
    LIST_PRODUCTS = 'list-products',
    DETAIL_PRODUCT = 'detail-product',
    CREATE_PRODUCT = 'create-product',
    EDIT_PRODUCT = 'edit-product',
    DELETE_PRODUCT = 'delete-product',

    // Clients
    LIST_CLIENTS = 'list-clients',
    DETAIL_CLIENT = 'detail-client',
    CREATE_CLIENT = 'create-client',
    EDIT_CLIENT = 'edit-client',
    DELETE_CLIENT = 'delete-client'
}